webscraper/requirements.txt
scrapy==2.11.2
elasticsearch==8.13.0
python-dotenv==1.0.1